rename SSIM file
