+++ torchimage/metrics/__init__.py
@@ -1,6 +1,6 @@
 from .psnr import PSNR
 from .mse import MSE
-from .ssim_new import SSIM, MS_SSIM
+from .ssim import SSIM, MS_SSIM
 
 
 __all__ = [
